requirements.txt
pandas==1.5.3
python-dotenv==1.0.0
ydata-profiling==v4.1.2
